CAMEL-16400: adjust manual integration tests for camel-ldap
