+++ components/camel-ldap/src/test/java/org/apache/camel/component/ldap/LdapRouteTwoManualIT.java
@@ -17,10 +17,9 @@
 package org.apache.camel.component.ldap;
 
 import java.util.Collection;
+import java.util.Hashtable;
 
-import javax.naming.directory.Attributes;
 import javax.naming.directory.SearchResult;
-import javax.naming.ldap.LdapContext;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.Endpoint;
@@ -29,6 +28,7 @@ import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.impl.DefaultCamelContext;
 import org.apache.camel.support.SimpleRegistry;
+import org.apache.directory.api.util.Network;
 import org.apache.directory.server.annotations.CreateLdapServer;
 import org.apache.directory.server.annotations.CreateTransport;
 import org.apache.directory.server.core.annotations.ApplyLdifFiles;
@@ -37,19 +37,19 @@ import org.apache.directory.server.core.integ5.DirectoryExtension;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import static org.apache.directory.server.integ.ServerIntegrationUtils.getWiredContext;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(DirectoryExtension.class)
 @CreateLdapServer(transports = { @CreateTransport(protocol = "LDAP") })
 @ApplyLdifFiles("org/apache/camel/component/ldap/LdapRouteTest.ldif")
-public class LdapRouteTest extends AbstractLdapTestUnit {
+@EnabledIfSystemProperty(named = "enable.ldap.itests", matches = "true",
+                         disabledReason = "the tests does not work due to complex ldap server environment")
+public class LdapRouteTwoManualIT extends AbstractLdapTestUnit {
 
     private CamelContext camel;
     private ProducerTemplate template;
@@ -60,10 +60,16 @@ public class LdapRouteTest extends AbstractLdapTestUnit {
         // you can assign port number in the @CreateTransport annotation
         port = super.getLdapServer().getPort();
 
-        LdapContext ctx = getWiredContext(ldapServer);
+        // use ENV variables
+        Hashtable<String, String> env = new Hashtable();
+        env.put("java.naming.factory.initial", "com.sun.jndi.ldap.LdapCtxFactory");
+        env.put("java.naming.provider.url", Network.ldapLoopbackUrl(ldapServer.getPort()));
+        env.put("java.naming.security.principal", "uid=admin,ou=system");
+        env.put("java.naming.security.credentials", "secret");
+        env.put("java.naming.security.authentication", "simple");
 
         SimpleRegistry reg = new SimpleRegistry();
-        reg.bind("localhost:" + port, ctx);
+        reg.bind("localhost:" + port, env);
         camel = new DefaultCamelContext(reg);
         template = camel.createProducerTemplate();
     }
@@ -74,11 +80,10 @@ public class LdapRouteTest extends AbstractLdapTestUnit {
     }
 
     @Test
-    public void testLdapRouteStandard() throws Exception {
+    public void testLdapRouteStandardTwo() throws Exception {
         camel.addRoutes(createRouteBuilder("ldap:localhost:" + port + "?base=ou=system"));
         camel.start();
 
-        // START SNIPPET: invoke
         Endpoint endpoint = camel.getEndpoint("direct:start");
         Exchange exchange = endpoint.createExchange();
         // then we set the LDAP filter on the in body
@@ -92,71 +97,21 @@ public class LdapRouteTest extends AbstractLdapTestUnit {
         assertTrue(contains("uid=test2,ou=test,ou=system", searchResults));
         assertTrue(contains("uid=testNoOU,ou=test,ou=system", searchResults));
         assertTrue(contains("uid=tcruise,ou=actors,ou=system", searchResults));
-        // START SNIPPET: invoke
-    }
 
-    @Test
-    public void testLdapRouteWithPaging() throws Exception {
-        camel.addRoutes(createRouteBuilder("ldap:localhost:" + port + "?base=ou=system&pageSize=5"));
-        camel.start();
-
-        Endpoint endpoint = camel.getEndpoint("direct:start");
-        Exchange exchange = endpoint.createExchange();
+        // call again
+        endpoint = camel.getEndpoint("direct:start");
+        exchange = endpoint.createExchange();
         // then we set the LDAP filter on the in body
-        exchange.getIn().setBody("(objectClass=*)");
+        exchange.getIn().setBody("(!(ou=test1))");
 
         // now we send the exchange to the endpoint, and receives the response from Camel
-        Exchange out = template.send(endpoint, exchange);
-
-        Collection<SearchResult> searchResults = defaultLdapModuleOutAssertions(out);
-        assertEquals(16, searchResults.size());
-    }
-
-    @Test
-    public void testLdapRouteReturnedAttributes() throws Exception {
-        // LDAP servers behave differently when it comes to what attributes are returned
-        // by default (Apache Directory server returns all attributes by default)
-        // Using the returnedAttributes parameter this behavior can be controlled
-
-        camel.addRoutes(createRouteBuilder("ldap:localhost:" + port + "?base=ou=system&returnedAttributes=uid,cn"));
-        camel.start();
-
-        Endpoint endpoint = camel.getEndpoint("direct:start");
-        Exchange exchange = endpoint.createExchange();
-        // then we set the LDAP filter on the in body
-        exchange.getIn().setBody("(uid=tcruise)");
-
-        // now we send the exchange to the endpoint, and receives the response from Camel
-        Exchange out = template.send(endpoint, exchange);
-        Collection<SearchResult> searchResults = defaultLdapModuleOutAssertions(out);
-
-        assertEquals(1, searchResults.size());
-        assertTrue(contains("uid=tcruise,ou=actors,ou=system", searchResults));
-        Attributes theOneResultAtts = searchResults.iterator().next().getAttributes();
-        assertEquals("tcruise", theOneResultAtts.get("uid").get());
-        assertEquals("Tom Cruise", theOneResultAtts.get("cn").get());
-
-        // make sure this att is NOT returned anymore 
-        assertNull(theOneResultAtts.get("sn"));
-    }
-
-    @Test
-    public void testLdapRoutePreserveHeader() throws Exception {
-        camel.addRoutes(createRouteBuilder("ldap:localhost:" + port + "?base=ou=system"));
-        camel.start();
-
-        Exchange out = template.request("direct:start", exchange -> {
-            exchange.getIn().setBody("(!(ou=test1))");
-            exchange.getIn().setHeader("ldapTest", "Camel");
-        });
-
-        Collection<SearchResult> searchResults = defaultLdapModuleOutAssertions(out);
+        out = template.send(endpoint, exchange);
+        searchResults = defaultLdapModuleOutAssertions(out);
 
         assertFalse(contains("uid=test1,ou=test,ou=system", searchResults));
         assertTrue(contains("uid=test2,ou=test,ou=system", searchResults));
         assertTrue(contains("uid=testNoOU,ou=test,ou=system", searchResults));
         assertTrue(contains("uid=tcruise,ou=actors,ou=system", searchResults));
-        assertEquals("Camel", out.getMessage().getHeader("ldapTest"));
     }
 
     @SuppressWarnings("unchecked")
